+++ crates/core/dep-graph/Cargo.toml
@@ -12,14 +12,12 @@ edition = "2021"
 
 [dependencies]
 moon_action = { path = "../action" }
-moon_config = { path = "../config" }
 moon_error = { path = "../error" }
 moon_logger = { path = "../logger" }
 moon_platform = { path = "../platform" }
 moon_project = { path = "../project" }
 moon_project_graph = { path = "../project-graph" }
 moon_task = { path = "../task" }
-moon_utils = { path = "../utils" }
 petgraph = { version = "0.6.2", features = ["serde"] }
 rustc-hash = { workspace = true }
 thiserror = { workspace = true }
