+++ crates/core/dep-graph/src/lib.rs
@@ -1,5 +1,7 @@
+mod dep_builder;
 mod dep_graph;
 mod errors;
 
-pub use dep_graph::DepGraph;
+pub use dep_builder::DepGraphBuilder;
+pub use dep_graph::*;
 pub use errors::DepGraphError;
